+++ contracts/interface/IToken.sol
@@ -4,12 +4,14 @@ pragma solidity ^0.8.0;
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
-interface IToken is IERC20{
-    function mint(address to, uint amount) external returns(bool);
-    function burn(address from, uint amount) external returns(bool);
+interface IToken is IERC20 {
+    function mint(address to, uint256 amount) external returns (bool);
+
+    function burn(address from, uint256 amount) external returns (bool);
 }
 
 interface IMB {
-    function mint(address to, uint amount) external returns(bool);
-    function burn(address from, uint amount) external returns(bool);
+    function mint(address to, uint256 amount) external returns (bool);
+
+    function burn(address from, uint256 amount) external returns (bool);
 }
